+++ database/migrations/2017_04_05_193324_add_is_gross_to_jobs_table.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddIsGrossToJobsTable extends Migration
+{
+    use SchemaBuilder;
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        $this->schema->table('jobs', function (Blueprint $table) {
+            $table->smallInteger('is_gross')->default(0);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        $this->schema->table('jobs', function (Blueprint $table) {
+            $table->dropColumn('is_gross');
+        });
+    }
+}